WIP: ItemView will overlay other components
The idea is that an ItemView component can "pop up" over any other "activity"
the user might be doing. This commit just adds the CSS to allow that. I still
have to add the stuff to make the ItemView appear (by URL) and disappear
(with an "X" in the top-right of the component that will switch out the URL).

Related to #33