Changed: TemplateSearch can collapse fields
The fields take a lot of space, and you have to scroll past them to see
your results. Now there's a button to show/hide the fields so you don't
have to do as much scrolling. This should be done automatically when
the user selects the "Search" button, but I'm not sure how yet.