Changed: imports and exports now explicit (they don't use "default")
This covers for a babel-jest issue where exports with "default" aren't mocked
in compliance with ES6. It's not clear whether this will ever be fixed, or
indeed if Facebook even thinks it's a problem... and it's not a big deal for
me either, so I'll just make our (ex|im)ports explicit.