Fixed: settting a SearchQuery Store field to zero-length string removes it
Previously, setting a field to a zero-length string would set it to a zero-
length string. Although CantusJS should prevent it from resulting in an
invalid query, it didn't (CANTUS-Project/cantus-js#16). Moreover, while the
solution in cantus-js#16 makes sense for CantusJS, it doesn't work in Vitrail.
An empty field here is just an empty field.