ItemView: add default values for many calls to Map.get()
When an ImmutableJS get() function can't "get" what you ask for (e.g., the
key doesn't exist in the Map) it returns "undefined" by default. This can
cause problems when get() is assumed to return a string, and for ItemView
it will always result in the word "undefined" appearing on screen.

Therefore, I've added default return values as the second argument to get(),
for the calls to get() where it's not already protected by an if(){}. The
default return value in this case is an empty string.