"log" module may now issue an alert()
Now, error and warning messages will result in a call to alert() so that
the user is aware something bad is happening. Otherwise it might look like
nothing happened, unless the user happend to have the browser console open.

Debug messages still won't be shown with alert() because I'm expecting people
to expect to have to look for them.

Finally, I know alert() isn't great, but it seems like the best way to guarantee
I won't end up with an endless cycle of error messages going through NuclearJS
and causing more error messages.