Refactor: ResultListTable has predefined field order for some types
If the results are all chants or all sources, there's a predefined column
order that's the same as the "primary" fields in the ItemView.

Note that this is only a temporary effort in my long-term attempt to refactor
the ResultListTable. The field order should really come from NuclearJS, but
this measure will help me troubleshoot Abbot in the meantime.

For #37