New: NuclearJS amendments so GUI knows when search is submitted
I added an extra field to the "SearchResults" Store, "loading," that can
be set to true when a server request is submitted, and that is always set
to false when search results are loaded. This will allow the GUI to show
a "loading" message when waiting for results from the server.

For #64